Improved cookie server and made it work for local servers
Removed global variables, and replaced with a "cookie server"
object instead. This also allows for customizations to how the
cookie server operates. For example, you can now change the
security of the cookie, to allow cookies to still work when
running the gameserver locally. (without the need for self-signed
certificates)